Implement a configurable request logger.
This PR creates a RequestLogger type which logs important request
information.

This type can be passed a specific logger (Syslog, StdOut etc.) to
which it writes.

The type is also passed a Formatter interface type, of which I have
implemented the LogStashFormatter attempting to mirror the current
Voice LogStash implementation. You can see a comparison here:

https://gist.github.com/ciaranarcher/f241539c5373e3e5e001

When testing I use a logger that wraps an in-memory buffer to
allow easy testing. For realz in the app I've included a simple
level-logger implementation you can find here:

http://godoc.org/github.com/tideland/goas/v2/logger